c to f

+++ celcius_to_fahrenheit.py
@@ -1,7 +1,7 @@
-celcius = int(input('Input Celsius temperature: '))
+celcius = float(input('Input Celsius temperature: '))
 fahrenheit = (celcius * 1.8) + 32
 print(fahrenheit)
 
-fahrenheit = int(input('Input Fahrenheit temperature: '))
+fahrenheit = float(input('Input Fahrenheit temperature: '))
 celsius = (fahrenheit - 32) / 1.8
 print(celsius)
